Turn all tests into integration tests (#495)
